requirements.txt
flask==3.0.3
flask-cors==4.0.1
langgraph==0.2.35
langchain==0.3.7
langchain-community==0.3.5
chromadb==0.5.5
sentence-transformers==3.2.1
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.8.3
python-dotenv==1.0.1
reportlab==4.2.5
gunicorn==23.0.0
pytest==8.3.3
flake8==7.1.1

